owenwowson/requirements.txt
discord.py
pyyaml
requests
PyNaCl